requirements.txt
matplotlib
scikit-learn
scikit-image==0.19.3 
opencv-python
yacs
joblib
natsort
h5py
tqdm
einops
gdown
addict
future
lmdb
numpy
pyyaml
requests
scipy
tb-nightly
yapf
lpips
torchmetrics